Remove overrides.zcml and instead override the viewlet by specifying a browserlayer.
Moved interfaces.py which contains the browserlayer one level up and renamed
according to convention.

This requires this product to be reinstalled on existing sites.

IThemeSpecific in browser/interfaces.py is no longer used, but I leave it for
now to not break existing sites that depend on it. (Reinstallation will remove
the dependency).